backend/requirements.txt
Django>=5.1.4
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
django-cors-headers==3.11.0
requests
prometheus_client==0.21.1
